BidirectionalAlgorithms

This section of the manual describes the module for bidirectional algorithms.

BidirectionalAlgorithms

NeuralNetworkReachability.BidirectionalAlgorithms.BidirectionalAlgorithm — Type

BidirectionalAlgorithm

Abstract supertype of bidirectional algorithms.

NeuralNetworkReachability.BidirectionalAlgorithms.SimpleBidirectional — Type

SimpleBidirectional{FA<:ForwardAlgorithm, BA<:BackwardAlgorithm} <: BidirectionalAlgorithm

Simple bidirectional algorithm parametric in a forward and backward algorithm.

Fields

fwd_algo – forward algorithm

bwd_algo – backward algorithm