ForwardAlgorithms

This section of the manual describes the module for forward algorithms.

ForwardAlgorithms

NeuralNetworkReachability.ForwardAlgorithms.ForwardAlgorithm — Type

ForwardAlgorithm

Abstract supertype of forward algorithms.

NeuralNetworkReachability.ForwardAlgorithms.DefaultForward — Type

DefaultForward <: ForwardAlgorithm

Default forward algorithm, which works for vector-like inputs.

NeuralNetworkReachability.ForwardAlgorithms.ConcreteForward — Type

ConcreteForward <: ForwardAlgorithm

Forward algorithm that uses concrete set operations.

NeuralNetworkReachability.ForwardAlgorithms.LazyForward — Type

LazyForward <: ForwardAlgorithm

Forward algorithm that uses lazy set operations.

NeuralNetworkReachability.ForwardAlgorithms.BoxForward — Type

BoxForward{AMA<:ForwardAlgorithm} <: ForwardAlgorithm

Forward algorithm that uses a box approximation for non-identity activations and applies the affine map according to the specified algorithm.

Fields

affine_map_algorithm – algorithm to apply for affine maps

NeuralNetworkReachability.ForwardAlgorithms.SplitForward — Type

SplitForward{S<:ForwardAlgorithm,FS,FM} <: ForwardAlgorithm

Forward algorithm that splits a set, then computes the image under the neural network, and finally merges the resulting sets again, all according to a policy.

Fields

algo – algorithm to be applied between splitting and merging

split_function – function for splitting

merge_function – function for merging

NeuralNetworkReachability.ForwardAlgorithms.DeepZ — Type

DeepZ <: ForwardAlgorithm

Forward algorithm based on zonotopes for ReLU, sigmoid, and tanh activation functions from [1].

[1]: Singh et al.: Fast and Effective Robustness Certification, NeurIPS 2018.

NeuralNetworkReachability.ForwardAlgorithms.Verisig — Type

Verisig{R} <: ForwardAlgorithm

Forward algorithm for sigmoid and tanh activation functions from [1].

Fields

algo – reachability algorithm of type TMJets

Notes

The implementation is known to be unsound in some cases.

The implementation currently only supports neural networks with a single hidden layer.

[1] Ivanov et al.: Verisig: verifying safety properties of hybrid systems with neural network controllers, HSCC 2019.